[restructure] cleaned up folder structure and started writing new early block allocator
