+++ bochs.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+. ./iso.sh
+
+bochs 'boot:cdrom' 'ata1-master: type=cdrom,path=missy.iso,status=inserted'
