+++ include/init/kerrno.h
@@ -1,6 +1,6 @@
 #ifndef _KERNEL_EARLY_ERRNO_H
 #define _KERNEL_EARLY_ERRNO_H
 
-extern int early_errno;
+extern int kinit_errno;
 
 #endif
