+++ include/init/kterm.h
