+++ include/init/kutils.h
