--- include/kstub_logger.h
@@ -1,6 +1,0 @@
-#ifndef _KSTUB_LOGGER_H
-#define _KSTUB_LOGGER_H
-
-
-
-#endif
